standardization of the calendar events (in progress)

+++ addons/calendar/static/src/views/list_view/calendar_list_controller.js
@@ -64,35 +64,47 @@ export class CaledarListController extends ListController {
 
     async onDeleteSelectedRecords() {
         const selectedRecords = this.model.root.selection;
-        if (selectedRecords.length == 1) {
+        const userAttendeesDetails = await this.orm.call("res.partner", "get_attendee_detail", [
+            user.partnerId,
+            selectedRecords.map((record) => record.resId),
+        ]);
+        if (selectedRecords.length == 1 && user.userId === record.data.user_id.id) {
+            console.log("is passing to delete only one record")
             const record = selectedRecords[0];
-            const userAttendeeDetail = await this.orm.call("res.partner", "get_attendee_detail", [
-                user.partnerId,
-                record.resIds,
-            ]);
-            const userAttendeeId = userAttendeeDetail[0].attendee_id;
-            if (user.userId === record.data.user_id.id) {
-                const partnerIds = record.data.partner_ids.resIds;
-                if (record.data.recurrency) {
-                    this.openRecurringDeletionWizard(record.resId, userAttendeeId);
-                } else if (partnerIds.length === 1 && partnerIds[0] === user.partnerId) {
-                    super.onDeleteSelectedRecords(...arguments);
-                } else {
-                    await this.orm.call("calendar.event", "action_unlink_event", [
-                        record.resIds,
-                        userAttendeeId,
-                    ])
-                    .then((action) => {
-                        if (action && action.context) {
-                            this.actionService.doAction(action);
-                        } else {
-                            location.reload();
-                        }
-                    });
-                }
+            const partnerIds = record.data.partner_ids.resIds;
+            const userAttendeeId = userAttendeesDetails[0].attendee_id;
+            if (record.data.recurrency) {
+                this.openRecurringDeletionWizard(record.resId, userAttendeeId);
+            } else if (partnerIds.length === 1 && partnerIds[0] === user.partnerId) {
+                super.onDeleteSelectedRecords(...arguments);
             } else {
-                // Decline event
-                this.orm.call("calendar.attendee", "do_decline", [[userAttendeeId]]);
+                await this.orm.call("calendar.event", "action_unlink_event", [
+                    record.resIds,
+                    userAttendeeId,
+                ])
+                .then((action) => {
+                    if (action && action.context) {
+                        this.actionService.doAction(action);
+                    } else {
+                        location.reload();
+                    }
+                });
+            }
+        } else {
+            const declinedAttendeeIds = [];
+            const deletedEventIds = [];
+            for (const userAttendeeDetails of userAttendeesDetails) {
+                if (userAttendeeDetails.is_organizer) {
+                    deletedEventIds.push(userAttendeeDetails.event_id);
+                } else {
+                    declinedAttendeeIds.push(userAttendeeDetails.attendee_id);
+                }
+            }
+            if (declinedAttendeeIds){
+                this.orm.call("calendar.attendee", "do_decline", [declinedAttendeeIds]);
+            }
+            if (deletedEventIds){
+                this.orm.call("calendar.event", "action_unlink_events", [deletedEventIds]);
             }
         }
     }
